% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-dat.R
\docType{data}
\name{dyad.sim}
\alias{dyad.sim}
\title{Simulated dyadic data from \code{\href{https://arxiv.org/pdf/1312.3398.pdf}{Aronow, Peter M., Cyrus Samii, and Valentina A. Assenova. "Cluster-robust variance estimation for dyadic data." Political Analysis 23.4 (2015): 564-577.}}}
\format{An object of class \code{data.frame} with 4950 rows and 5 columns.}
\source{
\href{https://doi.org/10.7910/DVN/OMJYE5}{Harvard Dataverse}
}
\usage{
data(dyad.sim)
}
\description{
Simulated data with 100 units and full pairwise dyads.
}
\examples{
data(dyad.sim)
head(dyad.sim)
}
\references{
Aronow, Peter M., Cyrus Samii, and Valentina A. Assenova. "Cluster-robust variance estimation for dyadic data." Political Analysis 23.4 (2015): 564-577.
(\href{https://arxiv.org/pdf/1312.3398.pdf}{ARXIV})
}
\keyword{datasets}
