% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyadRobust.R
\name{dyadRobust}
\alias{dyadRobust}
\title{dyadRobust}
\usage{
dyadRobust(fit, dat, dyadid, egoid, alterid)
}
\arguments{
\item{fit}{The model object. Should be of type 'lm'.}

\item{dat}{The dataset used to calculate the \code{fit} object.}

\item{dyadid}{Name of column containing the dyad ID. Must be a character.}

\item{egoid}{Name of column containing the ego ID. Must be a character.}

\item{alterid}{Name of column containing the alter ID. Must be a character.}
}
\value{
Returns a list containing the following components:
\item{bhat}{A \code{named vector} of coefficient estimates from the \code{fit} model object.}
\item{sehat}{A \code{named vector} of dyad-robust standard errors calculated via multiway decomposition.}
}
\description{
This function calculates dyad-robust standard errors via multiway decomposition, proposed by Aronow, Peter M., Cyrus Samii, and Valentina A. Assenova. "Cluster-robust variance estimation for dyadic data." Political Analysis 23.4 (2015): 564-577 (\code{\href{https://arxiv.org/pdf/1312.3398.pdf}{ARXIV}}).
}
\examples{
data("dyad.sim")
m <- lm(dY ~ dX,dyad.sim)
out <- dyadRobust(fit = m,
           dat = dyad.sim,
           dyadid = "dyads",
           egoid = "dyad1",
           alterid = "dyad2")
}
\keyword{cluster}
\keyword{dyad}
\keyword{dyadic}
\keyword{errors}
\keyword{inference}
\keyword{robust}
\keyword{standard}
